Increase test coverage
As of now, tests/master.py covers 76% of the functions in rrPython
and has the same approximate coverage as the old version. All functions
that are deliberately ignored either segfault or are annoying (enableLogging),
and the other functions either aren't tested by the old version or
had blatantly wrong tests.